% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_graphs.R
\name{assign_names}
\alias{assign_names}
\title{Assign names to nodes.}
\usage{
assign_names(model)
}
\arguments{
\item{model}{A trained \code{SICNN_Net} model with input-skip.}
}
\value{
A list of adjacency matrices with the correct names.
}
\description{
Internal helper function to assign descriptive names to nodes used for plotting.
}
\keyword{internal}
