% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_activation_functions.R
\name{Custom_activation}
\alias{Custom_activation}
\title{Generate a custom activation function.}
\usage{
Custom_activation()
}
\value{
Returns a `\code{torch::nn_module} that can be used in an \code{SICNN_Net}
}
\description{
The first 3 entries are customized
in order to see if we can learn that structure. The rest will be relu as usual
}
