% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_graphs.R
\name{get_adj_mats}
\alias{get_adj_mats}
\title{Obtain adjacency matrices for \code{igraph} plotting}
\usage{
get_adj_mats(model)
}
\arguments{
\item{model}{An instance of \code{SICNN_Net} with input_skip enabled.}
}
\value{
A list of adjacency matrices, one for each hidden layer and the output layer.
}
\description{
Takes the alpha active path matrices for each layer of the SICNN and converts
them to adjacency matrices so that they can be plotted with igraph.
}
\keyword{internal}
