% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_explanations.R
\name{get_fisher_information}
\alias{get_fisher_information}
\title{Compute parameter covariance for an SICNN model.}
\usage{
get_fisher_information(
  model,
  dataloader,
  type = c("diagonal", "block-diagonal", "KFAC"),
  device = "cpu"
)
}
\arguments{
\item{model}{A trained \code{SICNN_Net} object.}

\item{dataloader}{A \code{torch::dataloader} object.}

\item{type}{character, the type of approximation: \code{"diagonal"} or \code{"block-diagonal"}.}

\item{device}{character, the device to be used (default = \code{"cpu"}).}
}
\value{
A list containing the covariance matrix components (diagonal or blocks).
}
\description{
Computes the covariance matrix (or its approximation) for the weights of a trained SICNN model using the sandwich estimator (Fan and Li, 2001) for penalized models.
}
