% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite_functions.R
\name{get_input_active_counts}
\alias{get_input_active_counts}
\title{Count active input connections by layer}
\usage{
get_input_active_counts(model)
}
\arguments{
\item{model}{An instance of \code{SICNN_Net} where \code{input_skip = TRUE}.}
}
\value{
A matrix of shape (p, L-1) where p is the number of input variables
and L the total number of layers (including input and output), with each element being 1 if the variable is included
or 0 if not included.
}
\description{
Useful when the number of inputs and/or hidden neurons are very
large, and direct visualization of the network is difficult.
}
\keyword{internal}
