% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite_functions.R
\name{predict.SICNN_Net}
\alias{predict.SICNN_Net}
\title{Obtain deterministic predictions from an \code{SICNN} model}
\usage{
\method{predict}{SICNN_Net}(object, newdata, sparse = FALSE, device = "cpu", ...)
}
\arguments{
\item{object}{A trained \code{SICNN_Net} object.}

\item{newdata}{A \code{torch::dataloader} object containing the data with which to predict.}

\item{sparse}{logical, whether to use the thresholded active-path model.}

\item{device}{character, device for computation (default = \code{"cpu"}).}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A \code{torch::torch_tensor} of shape \code{(N,C)}.
}
\description{
Compute predictions from a trained \code{SICNN_Net} object.
}
