% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite_functions.R
\name{print.SICNN_Net}
\alias{print.SICNN_Net}
\title{Print summary of an \code{SICNN_Net} object}
\usage{
\method{print}{SICNN_Net}(x, ...)
}
\arguments{
\item{x}{An object of class \code{SICNN_Net}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the input \code{x}.
}
\description{
Provides a summary of a trained \code{SICNN_Net} object.
Includes the model type (input-skip or not), module and sub-module structure,
and number of trainable parameters.
}
\examples{
\donttest{
x <- torch::torch_randn(3, 2)
b <- torch::torch_rand(2)
y <- torch::torch_matmul(x, b)
train_data <- torch::tensor_dataset(x, y)
train_loader <- torch::dataloader(train_data, batch_size = 3, shuffle = FALSE)
problem <- "regression"
sizes <- c(2, 1, 1)
model <- SICNN_Net(problem, sizes, input_skip = TRUE)
print(model)
}
}
