% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SICNN-package.R
\docType{package}
\name{SICNN-package}
\alias{SICNN}
\alias{SICNN-package}
\title{SICNN: Smooth Information Criterion Neural Networks Using 'torch'}
\description{
Smooth Information Criterion Neural Networks (SICNNs) are implemented using 'torch', an R interface to the LibTorch backend. SICNNs provide a deterministic, frequentist approach to parameter pruning by leveraging a smooth approximation of the L0 norm (SIC), as described in O'Neill and Burke (2023) \doi{10.48550/arXiv.2212.12644}. The architecture supports deep structures with input-skip connections to facilitate feature selection in high-dimensional settings.
}
\author{
\strong{Maintainer}: Aliaksandr Hubin \email{aliaksandr.hubin@nmbu.no}

Authors:
\itemize{
  \item Andrew McInerney \email{andrew.mcinerney@ul.ie}
}

}
\keyword{internal}
