% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layers.R
\name{SICNN_Conv2d}
\alias{SICNN_Conv2d}
\title{Class to generate an SICNN convolutional layer}
\usage{
SICNN_Conv2d(
  in_channels,
  out_channels,
  kernel_size,
  stride = 1,
  padding = 0,
  dilation = 1,
  device = "cpu",
  bias = TRUE
)
}
\arguments{
\item{in_channels}{integer.}

\item{out_channels}{integer.}

\item{kernel_size}{integer or tuple.}

\item{stride}{integer or tuple.}

\item{padding}{integer, tuple or string.}

\item{dilation}{integer or tuple.}

\item{device}{The device to be used. Default is CPU.}

\item{bias}{logical.}
}
\description{
This module implements an SICNN Conv2d layer.
}
\keyword{internal}
