% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SICNN_Model.R
\name{SICNN_ConvNet}
\alias{SICNN_ConvNet}
\title{SICNN Convolutional Neural Network}
\usage{
SICNN_ConvNet(conv1, conv2, fc1, fc2, device = "cpu")
}
\arguments{
\item{conv1}{First convolutional layer (e.g. \code{SICNN_Conv2d}).}

\item{conv2}{Second convolutional layer.}

\item{fc1}{First fully connected layer (e.g. \code{SICNN_Linear}).}

\item{fc2}{Second fully connected layer (output layer).}

\item{device}{character, either 'cpu', 'gpu' or 'mps'.}
}
\value{
A \code{torch::nn_module} for Convolutional networks compatible with \code{train_SICNN}.
}
\description{
A Convolutional Network structured compatible with the SICNN framework.
}
