% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layers.R
\name{SICNN_Linear}
\alias{SICNN_Linear}
\title{Class to generate an SICNN feed forward layer}
\usage{
SICNN_Linear(
  in_features,
  out_features,
  device = "cpu",
  bias = TRUE,
  conv_net = FALSE
)
}
\arguments{
\item{in_features}{integer, number of input neurons.}

\item{out_features}{integer, number of output neurons.}

\item{device}{The device to be used. Default is CPU.}

\item{bias}{logical, whether to use bias.}
}
\description{
This module implements a fully connected SICNN layer.
}
\keyword{internal}
