% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_graphs.R
\name{SICNN_plot}
\alias{SICNN_plot}
\title{Function to plot an input skip structure after removing weights in non-active paths.}
\usage{
SICNN_plot(
  model,
  layer_spacing = 1,
  neuron_spacing = 1,
  vertex_size = 10,
  label_size = 0.5,
  edge_width = 0.5,
  save_svg = NULL
)
}
\arguments{
\item{model}{A trained \code{SICNN_Net} model with input_skip enabled.}

\item{layer_spacing}{numeric, spacing in between layers.}

\item{neuron_spacing}{numeric, spacing between neurons within a layer.}

\item{vertex_size}{numeric, size of the neurons.}

\item{label_size}{numeric, size of the text within neurons.}

\item{edge_width}{numeric, width of the edges connecting neurons.}

\item{save_svg}{the path where the plot will be saved if save_svg is not null.}
}
\value{
This function produces plots as a side effect and does not return a value.
}
\description{
Uses igraph to plot.
}
\examples{
\donttest{
sizes <- c(2,3,3,1)
problem <- 'regression'
device <- 'cpu'
torch::torch_manual_seed(0)
model <- SICNN_Net(problem_type = problem, sizes = sizes,
                   input_skip = TRUE, device = device)
x <- torch::torch_randn(3,2)
b <- torch::torch_rand(2)
y <- torch::torch_matmul(x,b)$unsqueeze(2)
train_data <- torch::tensor_dataset(x,y)
train_loader <- torch::dataloader(train_data, batch_size = 3)
train_SICNN(epochs = 1, SICNN = model, lr = 0.01, train_dl = train_loader, n_train = 3)
SICNN_plot(model, 1, 1, 14, 1)
}
}
