% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite_functions.R
\name{summary.SICNN_Net}
\alias{summary.SICNN_Net}
\title{Summary of SICNN fit}
\usage{
\method{summary}{SICNN_Net}(object, epsilon = 1e-05, threshold = 0.5, ...)
}
\arguments{
\item{object}{An object of class \code{SICNN_Net}.}

\item{epsilon}{numeric, the final epsilon parameter corresponding to the trained model. Default is 1e-5.}

\item{threshold}{numeric, the threshold for determining active edges. Default is 0.5.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A \code{data.frame} containing the above information. The function prints a formatted summary to the console.
The returned \code{data.frame} is invisible.
}
\description{
Summary method for objects of the \code{SICNN_Net} class.
Only applies to objects trained with \code{input_skip = TRUE}.
}
\details{
The returned table combines two types of information:
\itemize{
\item Number of active edges from each input variable in each layer.
\item Proportion of active edges from each input variable in each layer, including a final
column showing the average across all layers.
}
}
\examples{
\donttest{
x <- torch::torch_randn(3, 2)
b <- torch::torch_rand(2)
y <- torch::torch_matmul(x, b)
train_data <- torch::tensor_dataset(x, y)
train_loader <- torch::dataloader(train_data, batch_size = 3, shuffle = FALSE)
problem <- "regression"
sizes <- c(2, 1, 1)
model <- SICNN_Net(problem, sizes, input_skip = TRUE)
train_SICNN(epochs = 1, SICNN = model, lr = 0.01, train_dl = train_loader, n_train = 3)
summary(model)
}
}
