% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train_validate.R
\name{validate_SICNN}
\alias{validate_SICNN}
\title{Validate a trained SICNN model.}
\usage{
validate_SICNN(SICNN, test_dl, device = "cpu", ...)
}
\arguments{
\item{SICNN}{An instance of a trained \code{SICNN_Net} to be validated.}

\item{test_dl}{A \code{torch::dataloader} to evaluate the model on.}

\item{device}{character, either 'cpu', 'gpu' or 'mps'.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list containing the following elements:
\describe{
\item{accuracy_full_model}{Classification accuracy of the full (dense) model (if classification).}
\item{accuracy_sparse}{Classification accuracy using only weights in active paths (if classification).}
\item{validation_error}{Root mean squared error for the full model (if regression).}
\item{validation_error_sparse}{Root mean squared error using only weights in active paths (if regression).}
\item{density}{Proportion of weights classified as active by the SIC threshold rule.}
\item{density_active_path}{Proportion of weights classified as active after removing edges outside active paths.}
}
}
\description{
Computes metrics on a validation dataset without computing gradients.
Returns metrics for both the full model and the sparse model.
}
