% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_explanations.R
\name{get_local_explanations_gradient}
\alias{get_local_explanations_gradient}
\title{Function to get gradient based local explanations for input-skip SICNNs.}
\usage{
get_local_explanations_gradient(
  model,
  input_data,
  magnitude = TRUE,
  include_potential_contribution = FALSE,
  device = "cpu",
  uncertainty = FALSE,
  fisher_dataloader = NULL,
  covariance_type = c("diagonal", "block-diagonal", "KFAC"),
  use_pseudo_inverse = FALSE,
  fisher = NULL,
  ...
)
}
\arguments{
\item{model}{A \code{SICNN_Net} with input-skip}

\item{input_data}{The data to be explained (one sample).}

\item{magnitude}{If TRUE, only return explanations. If FALSE, multiply by input values.}

\item{include_potential_contribution}{IF TRUE, If covariate=0,
we assume that the contribution is negative (good/bad that it is not included)
if FALSE, just removes zero covariates.}

\item{device}{character, the device to be trained on. Default is 'cpu', can be 'mps' or 'gpu'.}

\item{uncertainty}{logical, whether to compute uncertainty using the Delta method.}

\item{fisher_dataloader}{A \code{torch::dataloader} to compute the Fisher Information. If NULL and uncertainty is TRUE, it defaults to training data if available.}

\item{covariance_type}{character, the type of Fisher Information approximation: \code{"diagonal"}, \code{"block-diagonal"}, or \code{"KFAC"}.}

\item{use_pseudo_inverse}{logical, whether to use the pseudo-inverse if the Fisher Information is singular.}

\item{...}{further arguments (e.g., deprecated num_samples periodically used in old scripts).}
}
\value{
A list with the following elements:
\describe{
\item{explanations}{A \code{torch::tensor} of shape (1, p, num_classes).}
\item{p}{integer, the number of input features.}
\item{predictions}{A \code{torch::tensor} of shape (1, num_classes).}
\item{se}{A \code{torch::tensor} of shape (p, num_classes) containing standard errors (if uncertainty=TRUE).}
}
}
\description{
Works by computing the gradient wrt to input, given we have
relu activation functions.
}
