% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_graphs.R
\name{plot.SICNN_ConvNet}
\alias{plot.SICNN_ConvNet}
\title{Plot SICNN Convolutional Network}
\usage{
\method{plot}{SICNN_ConvNet}(x, ...)
}
\arguments{
\item{x}{An object of class \code{SICNN_ConvNet}}

\item{...}{Additional arguments.}
}
\value{
Invisible NULL.
}
\description{
Convolutional Networks cannot be plotted natively using the graph structures designed for Fully Connected Layers.
}
