% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_explanations.R
\name{plot_local_explanations_gradient}
\alias{plot_local_explanations_gradient}
\title{Plot the gradient based local explanations for one sample with input-skip SICNNs.}
\usage{
plot_local_explanations_gradient(
  model,
  input_data,
  device = "cpu",
  uncertainty = FALSE,
  fisher_dataloader = NULL,
  covariance_type = c("diagonal", "block-diagonal", "KFAC"),
  use_pseudo_inverse = FALSE,
  save_svg = NULL,
  ...
)
}
\arguments{
\item{model}{An instance of \code{SICNN_Net} with input-skip enabled.}

\item{input_data}{The data to be explained (one sample).}

\item{device}{character, the device to be trained on. Default is cpu. Can be 'mps' or 'gpu'.}

\item{uncertainty}{logical, whether to include Delta method uncertainty.}

\item{fisher_dataloader}{A \code{torch::dataloader} to compute the Fisher Information.}

\item{covariance_type}{character, the type of Fisher Information approximation: \code{"diagonal"}, \code{"block-diagonal"}, or \code{"KFAC"}.}

\item{use_pseudo_inverse}{logical, whether to use the pseudo-inverse if the Fisher Information is singular.}

\item{save_svg}{the path where the plot will be saved as svg, if save_svg is not NULL.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
This function produces plots as a side effect and does not return a value.
}
\description{
Plots the contribution of each covariate, and the prediction, with error bars.
}
