% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SICNN-package.R
\docType{package}
\name{SICNN-package}
\alias{SICNN}
\alias{SICNN-package}
\title{SICNN: Latent Binary Bayesian Neural Networks Using 'torch'}
\description{
Latent binary Bayesian neural networks (SICNNs) are implemented using 'torch', an R interface to the LibTorch backend. Supports mean-field variational inference as well as flexible variational posteriors using normalizing flows. The standard SICNN implementation follows Hubin and Storvik (2024) \doi{10.3390/math12060788}, using the local reparametrization trick as in Skaaret-Lund et al. (2024) \url{https://openreview.net/pdf?id=d6kqUKzG3V}. Input-skip connections are also supported, as described in Høyheim et al. (2025) \doi{10.48550/arXiv.2503.10496}.
}
\author{
\strong{Maintainer}: Lars Skaaret-Lund \email{lars.skaaret-lund@nmbu.no}

Authors:
\itemize{
  \item Aliaksandr Hubin \email{aliaksandr.hubin@nmbu.no}
  \item Eirik Høyheim \email{eirik.hoyheim@ffi.no}
}

}
\keyword{internal}
