% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SICNN_Model.R
\name{SICNN_Net}
\alias{SICNN_Net}
\title{Feed-forward Latent Binary Bayesian Neural Network (SICNN)}
\usage{
SICNN_Net(
  problem_type,
  sizes,
  input_skip = FALSE,
  device = "cpu",
  raw_output = FALSE,
  custom_act = NULL,
  link = NULL,
  nll = NULL,
  bias = TRUE
)
}
\arguments{
\item{problem_type}{character, one of:
\code{'binary classification'}, \code{'multiclass classification'}, \code{'regression'}, or \code{'custom'}.}

\item{sizes}{Integer vector specifying the layer sizes of the network. The first element is the input size,
the last is the output size, and the intermediate integers represent hidden layers.}

\item{input_skip}{logical, whether to include input_skip.}

\item{device}{the device to be trained on. Can be 'cpu', 'gpu' or 'mps'. Default is cpu.}

\item{raw_output}{logical, whether the network skips the last sigmoid/softmax layer to compute local explanations.}

\item{custom_act}{Allows the user to submit their own customized activation function.}

\item{link}{User can define their own link function (not implemented yet).}

\item{nll}{User can define their own likelihood function (not implemented yet).}

\item{bias}{logical, whether to include bias terms in linear layers. Default is TRUE.}
}
\value{
A \code{torch::nn_module} object representing the SICNN.
It includes the following methods:
\itemize{
\item \code{forward(x, sparse = FALSE)}: Performs a forward pass through the whole network.
\item \code{kl_div()}: Returns the KL divergence of the network.
\item \code{density()}: Returns the density of the whole network, i.e. the proportion of weights
with inclusion probabilities greater than 0.5.
\item \code{compute_paths()}: Computes active paths through the network without input-skip.
\item \code{compute_paths_input_skip()}: Computes active paths with  input-skip enabled.
\item \code{density_active_path()}: Returns network density after removing inactive paths.
\item \code{smooth_param_count(epsilon)}: Returns the smooth L0-based effective parameter
count used in the smooth information criterion (SIC) of O’Neill and Burke (2023),
based on the penalty \eqn{\phi_\epsilon(x) = x^2 / (x^2 + \epsilon^2)} applied to the
layer weight means.
}
}
\description{
Each layer is defined by \code{SICNN_Linear}.
For example, \code{sizes = c(20, 200, 200, 5)} generates a network with:
\itemize{
\item 20 input features,
\item two hidden layers of 200 neurons each,
\item an output layer with 5 neurons.
}
}
\examples{
\donttest{
layers <- c(10,2,5) 
prob <- 'multiclass classification'
net <- SICNN_Net(problem_type = prob, sizes = layers, input_skip = FALSE, device = 'cpu')
x <- torch::torch_rand(20,10,requires_grad = FALSE)
output <- net(x) 
net$sic_density(epsilon=1e-5, threshold=0.5) 
}
}
